Extract shared codex-payload helpers, fix dashboard leak in test
- Extract getKey/getMapKey/mapPath into src/domain/codex-payload.ts
  as the single canonical implementation for traversing untyped Codex
  JSON payloads with snake_case/camelCase key tolerance. Both
  running-entry.ts and tui.ts now import from this shared module,
  eliminating divergent null semantics (running-entry used ?? which
  treats explicit null as missing; codex-payload uses Object.hasOwn).
- Wrap e2e test runOnce() in try/finally so dashboard.stop() runs
  even if the orchestrator throws, preventing timer leaks across tests.
- rateLimits scaffolding is intentional — population deferred to a
  follow-up (noted in PR).
